Refactoring of opened editors list sample plugin
