--- samples/opened-editor-list/README.md
@@ -1,2 +1,0 @@
-# opened-editor-list
-opened-editor-list Plugin example for Theia.
