+++ samples/opened-editors-list/.gitignore
