+++ samples/opened-editors-list/README.md
@@ -1,0 +1,2 @@
+# opened-editors-list
+opened-editors-list Plugin example for Theia.
