+++ samples/opened-editors-list/tsconfig.json
